+++ include/Shape2D.hh
@@ -5,9 +5,9 @@ namespace Geometry {
 
     class Shape2D {
     public:
-        virtual double getArea() const = 0;
-        virtual double getPerimeter() const = 0;
-        virtual void move(double dx, double dy) = 0;
+        virtual float getArea() const = 0;
+        virtual float getPerimeter() const = 0;
+        virtual void move(float dx, float dy) = 0;
         virtual ~Shape2D() = default;
     };
